# contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "coursebook.golf API",
    "version": "1.0.0",
    "description": "JSON API for the coursebook.golf web, iOS and Android clients. Authenticate with a Clerk session token: `Authorization: Bearer <token>`. Every failure returns the ApiError envelope. The contract only grows: fields and error codes are added, never renamed or removed."
  },
  "servers": [
    {
      "url": "https://api.coursebook.golf"
    }
  ],
  "components": {
    "securitySchemes": {
      "ClerkSession": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A Clerk session token from the web, iOS or Android SDK."
      }
    },
    "schemas": {
      "Me": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "displayName"
        ]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ]
      },
      "ApiErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable machine-readable code. Current values: bad_request, validation_failed, us_state_required, unauthenticated, account_deleted, username_invalid, not_found, course_not_found, not_on_list, round_not_found, member_not_found, friendship_not_found, payload_too_large, unsupported_media_type, account_deletion_incomplete, search_unavailable, internal. New codes may appear; treat unknown codes by HTTP status.",
            "example": "not_on_list"
          },
          "message": {
            "type": "string",
            "description": "Explanation suitable to show to the member."
          },
          "requestId": {
            "type": "string",
            "description": "Correlates the failure with server logs."
          },
          "fields": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "Per-field problems for validation_failed; otherwise null."
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "fields"
        ]
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "description": "Dotted path of the invalid input, e.g. `q` or `course.name`."
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "message"
        ]
      },
...
      "CourseRanks": {
        "type": "object",
        "properties": {
          "world": {
            "type": "integer",
            "nullable": true,
            "minimum": 1
          },
          "usa": {
            "type": "integer",
            "nullable": true,
            "minimum": 1
          },
          "usaPublic": {
            "type": "integer",
            "nullable": true,
            "minimum": 1
          },
          "state": {
            "type": "integer",
            "nullable": true,
            "minimum": 1,
            "description": "Rank on the Best-in-State list of the course's own state."
          }
        },
        "required": [
          "world",
          "usa",
          "usaPublic",
          "state"
        ]
      },
...
      "SearchResults": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchHit"
            },
            "description": "At most 10, best match first."
          }
        },
        "required": [
          "results"
        ]
      },
      "SearchHit": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Catalog course id when the hit is already in the catalog; null otherwise."
          },
          "course": {
            "$ref": "#/components/schemas/CourseDetails"
          },
          "ranks": {
            "$ref": "#/components/schemas/CourseRanks"
          }
        },
        "required": [
          "courseId",
          "course",
          "ranks"
        ]
      },
      "CourseDetails": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "websiteUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "location",
          "city",
          "state",
          "country",
          "logoUrl",
          "websiteUrl"
        ],
        "description": "A course described by its details, e.g. a search result not yet in the catalog."
      },
      "Members": {
        "type": "object",
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true,
            "description": "Pass as `cursor` for the next page; null on the last page."
          }
        },
        "required": [
          "members",
          "nextCursor"
        ]
      },
      "Member": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "displayName"
        ],
        "description": "Another member's public projection; never an email or an id."
      },
...
      "MemberSearchResults": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberRelationship"
            },
            "description": "At most 20, usernames starting with the query first."
          }
        },
        "required": [
          "results"
        ]
      },
      "MemberRelationship": {
        "type": "object",
        "properties": {
          "member": {
            "$ref": "#/components/schemas/Member"
          },
          "relationship": {
            "type": "string",
            "description": "The viewer's relationship: none, friends, requested (the viewer asked; waiting for them) or incoming (they asked; the viewer can accept or decline). New values may appear."
          }
        },
        "required": [
          "member",
          "relationship"
        ]
      },
...
    },
...
  },
  "paths": {
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "tags": [
          "Account"
        ],
        "summary": "The signed-in member",
        "security": [
          {
            "ClerkSession": []
          }
        ],
        "responses": {
          "200": {
            "description": "The member.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: `unauthenticated`, or `account_deleted` after the account was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The Clerk account cannot use the app: `username_invalid`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteMe",
        "tags": [
          "Account"
        ],
        "summary": "Delete the member's account",
        "description": "Deletes the member's list and rounds, removes their name and email, and deletes the sign-in account. Courses they added stay in the shared catalog. Safe to retry: after a 502, call it again with the same token.",
        "security": [
          {
            "ClerkSession": []
          }
        ],
        "responses": {
          "204": {
            "description": "The account is deleted; sign the member out."
          },
          "401": {
            "description": "No valid session: `unauthenticated`, or `account_deleted` after the account was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The Clerk account cannot use the app: `username_invalid`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "The data is deleted but the sign-in account could not be: `account_deletion_incomplete`. Retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/course-search": {
      "get": {
        "operationId": "searchCourses",
        "tags": [
          "Catalog"
        ],
        "summary": "Find courses by name",
        "description": "Catalog courses carry their id; other hits carry details to send back when adding the course.",
        "security": [
          {
            "ClerkSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 2,
              "description": "Course name, at least two characters."
            },
            "required": true,
            "description": "Course name, at least two characters.",
            "name": "q",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Up to ten hits, best first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResults"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters: `validation_failed`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: `unauthenticated`, or `account_deleted` after the account was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The Clerk account cannot use the app: `username_invalid`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Course discovery is unavailable: `search_unavailable`. Retry later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/members": {
      "get": {
        "operationId": "listMembers",
        "tags": [
          "Members"
        ],
        "summary": "The member's friends, a page at a time",
        "description": "Only accepted friends are listed; pending requests are in `listFriendRequests`.",
        "security": [
          {
            "ClerkSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "nullable": true
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "nullable": true,
              "minimum": 1,
              "maximum": 200,
              "description": "Defaults to 50."
            },
            "required": false,
            "description": "Defaults to 50.",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "One page of members.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Members"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters: `validation_failed`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: `unauthenticated`, or `account_deleted` after the account was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The Clerk account cannot use the app: `username_invalid`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/member-search": {
      "get": {
        "operationId": "searchMembers",
        "tags": [
          "Members"
        ],
        "summary": "Find members by username to send a friend request",
        "description": "Returns usernames, display names and the viewer's relationship; never a list.",
        "security": [
          {
            "ClerkSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 3,
              "maxLength": 64,
              "description": "Text in the username, at least three characters."
            },
            "required": true,
            "description": "Text in the username, at least three characters.",
            "name": "q",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Matching members.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberSearchResults"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters: `validation_failed`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No valid session: `unauthenticated`, or `account_deleted` after the account was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "The Clerk account cannot use the app: `username_invalid`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  }
}